Re-arrange files back to storeConfig now we don't need to support...
..an exported storeConfig
